+++ lib/clixon/clixon_xml_map.h
@@ -62,10 +62,6 @@ int xml_namespace_change(cxobj *x, char *namespace, char *prefix);
 int xml_default(cxobj *x, void  *arg);
 int xml_sanity(cxobj *x, void  *arg);
 int xml_non_config_data(cxobj *xt, void *arg);
-int xml_bind_yang_rpc(cxobj *xrpc, yang_stmt *yspec, cxobj **xerr);
-int xml_bind_yang_rpc_reply(cxobj *xrpc, char *name, yang_stmt *yspec, cxobj **xerr);
-int xml_bind_yang0(cxobj *xt, yang_bind yb, yang_stmt *yspec, cxobj **xerr);
-int xml_bind_yang(cxobj *xt, yang_bind yb, yang_stmt *yspec, cxobj **xerr);
 
 int xml2xpath(cxobj *x, char **xpath);
 int assign_namespaces(cxobj *x0, cxobj *x1, cxobj *x1p);